Rename bokeh.util.{compiling->compiler}
