+++ bokeh/util/compiler.py
